functions/requirements.txt
pyTelegramBotAPI
google-generativeai
openai-whisper
python-dotenv
Flask
firebase-functions
firebase-admin